requirements.txt
google-generativeai
neo4j
sentence-transformers
faiss-cpu
streamlit
pandas